Update shopify_cache.rb
There are no ids for each object. The original developer took the straight object definition from Shopify such as collection:

{"collection":{"id":168742486074,"handle":"viper-vibe-one-time-purchase","updated_at":"2021-02-22T11:54:25-08:00","published_at":"2021-02-22T11:53:27-08:00","sort_order":"best-selling","template_suffix":"","products_count":3,"collection_type":"custom","published_scope":"web","title":"Viper Vibe - One Time Purchase","body_html":""}}

https://ellieactive.myshopify.com/admin/collections/168742486074.json

That is the Viper Vibe - 3 Items collection -- and used the ID NOT AS THE ACTIVE RECORD ID, but the id of the actual collection. Thus in the Database schema he created the id for the collection is the collection_id. That way he could take the JSON response and just dump it into the database. 

The current code base does not work as nothing is connected. So we can't push orders as the id used to connect objects to each other is instead filled with the Active Record id. Can you please revert to the pattern the original developer used?